+++ packages/progress/src/components/Spinner/Loader.tsx
@@ -1,6 +1,6 @@
 import {Loader as LoaderSVG, useTheme} from '@npm/mobydick-styles';
 import React, {FC} from 'react';
-import {Color} from 'react-native-svg';
+import {FillProps} from 'react-native-svg';
 
 import {ISizeSpinner, LoaderProps} from './types';
 
@@ -8,7 +8,7 @@ const Loader: FC<LoaderProps> = props => {
   const {size, fill, ...otherProps} = props;
   const {colors} = useTheme();
 
-  const getSize = (color: Color) => {
+  const getSize = (color: NonNullable<FillProps['fill']>) => {
     switch (size) {
       case ISizeSpinner.XXS:
         return (